# components/api-server/pkg/api/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  description: HyperShell gateway management API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: HyperShell API
  version: 1.0.0
servers:
- description: localhost
  url: http://localhost:8000
- description: current domain
  url: /
security:
- Bearer: []
...
paths:
  /api/hypershell/v1/gateways/deletion:
    get:
      operationId: getGatewayDeletionStatus
      parameters:
      - in: query
        name: external_reference
        required: true
        schema:
          maxLength: 255
          minLength: 1
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDeletionStatus"
          description: Gateway deletion state
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No reference exists in the caller scope
      security:
      - Bearer: []
      summary: Get deletion status for a reference in the caller scope
...
  /api/hypershell/v1/gateways:
    get:
      operationId: listGateways
      parameters:
      - description: Page number of record list when record list exceeds specified
          page size
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Maximum number of records to return
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 100
          minimum: 0
          type: integer
        style: form
      - description: Specifies the search criteria
        explode: true
        in: query
        name: search
        required: false
        schema:
          type: string
        style: form
      - description: Specifies the order by criteria
        explode: true
        in: query
        name: orderBy
        required: false
        schema:
          type: string
        style: form
      - description: Supplies a comma-separated list of fields to be returned
        explode: true
        in: query
        name: fields
        required: false
        schema:
          type: string
        style: form
      - description: Exact reference lookup scoped to the authenticated creator. Normal
          gateway access checks apply.
        in: query
        name: external_reference
        schema:
          maxLength: 255
          minLength: 1
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayList"
          description: A JSON array of gateway objects
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error occurred
      security:
      - Bearer: []
      summary: Returns a list of gateways
    post:
      description: Repeated calls with the same external_reference return the original
        gateway without updates. A deleted reference returns 409. Omit the field for
        independent creation.
      operationId: createGateway
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewayCreateRequest"
        description: Gateway data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation errors occurred
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gateway already exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: An unexpected error occurred creating the gateway
      security:
      - Bearer: []
      summary: Create a gateway or return the gateway for the caller and external
        reference
...
  /api/hypershell/v1/gateways/{gateway_id}/service_accounts:
    get:
      operationId: listGatewayServiceAccounts
      parameters:
      - description: Selected Gateway ID
        in: path
        name: gateway_id
        required: true
        schema:
          type: string
      - description: Page number of record list when record list exceeds specified
          page size
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Maximum number of records to return
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 100
          minimum: 0
          type: integer
        style: form
      - in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/OpenShellGatewayServiceAccountStatus"
      - description: Specifies the search criteria
        explode: true
        in: query
        name: search
        required: false
        schema:
          type: string
        style: form
      - in: query
        name: sort
        required: false
        schema:
          default: created_at
          enum:
          - name
          - role
          - status
          - expires_at
          - created_at
          type: string
      - in: query
        name: order
        required: false
        schema:
          default: desc
          enum:
          - asc
          - desc
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenShellGatewayServiceAccountList"
          description: Visible service accounts and caller capabilities
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid request or query value
        "401":
          description: Access token is missing or invalid
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gateway or service account was not found or is not visible
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error occurred
      summary: List OpenShell gateway service accounts
    post:
      operationId: createGatewayServiceAccount
      parameters:
      - description: Selected Gateway ID
        in: path
        name: gateway_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenShellGatewayServiceAccountCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenShellGatewayServiceAccountCreateResponse"
          description: Service account created; the credential is returned once
          headers:
            Cache-Control:
              description: Prevent storage of the one-time credential response
              schema:
                example: no-store
                type: string
            Pragma:
              description: Prevent legacy caches from storing the response
              schema:
                example: no-cache
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid request or query value
        "401":
          description: Access token is missing or invalid
        "403":
          description: Access token does not have sufficient privileges
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gateway or service account was not found or is not visible
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gateway is not ready or the active name already exists
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: A service-account quota is exhausted
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Keycloak provisioning or lifecycle verification is unavailable
      summary: Create an OpenShell gateway service account
...
  /api/hypershell/v1/roles:
    get:
      operationId: listRoles
      parameters:
      - description: Page number of record list when record list exceeds specified
          page size
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Maximum number of records to return
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 100
          minimum: 0
          type: integer
        style: form
      - description: Specifies the search criteria
        explode: true
        in: query
        name: search
        required: false
        schema:
          type: string
        style: form
      - description: Specifies the order by criteria
        explode: true
        in: query
        name: orderBy
        required: false
        schema:
          type: string
        style: form
      - description: Supplies a comma-separated list of fields to be returned
        explode: true
        in: query
        name: fields
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleList"
          description: A list of roles
        "401":
          description: Access token is missing or invalid
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error occurred
      summary: List all roles
...
components:
...
  schemas:
    List:
      allOf:
      - properties:
          kind:
            type: string
          page:
            type: integer
          size:
            type: integer
          total:
            type: integer
        type: object
      - $ref: "#/components/schemas/ObjectReference"
    ObjectReference:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        href: href
      properties:
        id:
          type: string
        kind:
          type: string
        href:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object
    Error:
      allOf:
      - $ref: "#/components/schemas/ObjectReference"
      - properties:
          code:
            type: string
          reason:
            type: string
          operation_id:
            type: string
        type: object
      example:
        reason: reason
        code: code
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        operation_id: operation_id
        id: id
        href: href
...
    Gateway:
      allOf:
      - $ref: "#/components/schemas/ObjectReference"
      - properties:
          name:
            type: string
          cluster_id:
            type: string
          release_id:
            type: string
          database_id:
            description: Server-assigned ManagedDatabase identifier; client-supplied
              values are ignored
            type: string
          namespace:
            description: API-assigned Kubernetes namespace derived from the Gateway
              identifier
            readOnly: true
            type: string
          external_reference:
            description: Immutable external resource reference scoped to its authenticated
              creator
            maxLength: 255
            minLength: 1
            type: string
          external_dns:
            type: string
          tls_mode:
            type: string
          service_type:
            type: string
          status:
            type: string
          phase:
            type: string
          image:
            description: Container image for the gateway deployment
            type: string
          supervisor_image:
            description: Container image for the supervisor sidecar
            type: string
          server_dns_names:
            description: DNS names for TLS certificate SANs
            items:
              type: string
            type: array
          route_address:
            description: External route address populated by the control plane
            readOnly: true
            type: string
          console_address:
            description: Web console address populated by the control plane
            readOnly: true
            type: string
          oidc:
            description: JSON-encoded OIDC authentication configuration (auto-populated
              by Keycloak provisioning)
            readOnly: true
            type: string
          route:
            description: JSON-encoded route configuration
            type: string
          credential_driver:
            description: JSON-encoded credential storage driver configuration
            type: string
          active_sandbox_count:
            description: Number of active (Running or Pending) agent sandboxes observed
              in the gateway namespace by the control plane
            format: int32
            readOnly: true
            type: integer
          created_by:
            description: "Username of the user who provisioned this gateway, resolved\
              \ from RBAC role bindings"
            readOnly: true
            type: string
        required:
        - cluster_id
        - database_id
        - name
        - namespace
        - release_id
        type: object
      example:
        release_id: release_id
        server_dns_names:
        - server_dns_names
        - server_dns_names
        created_at: 2000-01-23T04:56:07.000+00:00
        credential_driver: credential_driver
        cluster_id: cluster_id
        updated_at: 2000-01-23T04:56:07.000+00:00
        external_reference: external_reference
        id: id
        href: href
        external_dns: external_dns
        phase: phase
        image: image
        active_sandbox_count: 5
        kind: kind
        console_address: console_address
        route_address: route_address
        oidc: oidc
        created_by: created_by
        supervisor_image: supervisor_image
        service_type: service_type
        tls_mode: tls_mode
        route: route
        database_id: database_id
        name: name
        namespace: namespace
        status: status
    GatewayList:
      allOf:
      - $ref: "#/components/schemas/List"
      - properties:
          items:
            items:
              $ref: "#/components/schemas/Gateway"
            type: array
        type: object
      example:
        total: 1
        size: 6
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        page: 0
        id: id
        href: href
        items:
        - release_id: release_id
          server_dns_names:
          - server_dns_names
          - server_dns_names
          created_at: 2000-01-23T04:56:07.000+00:00
          credential_driver: credential_driver
          cluster_id: cluster_id
          updated_at: 2000-01-23T04:56:07.000+00:00
          external_reference: external_reference
          id: id
          href: href
          external_dns: external_dns
          phase: phase
          image: image
          active_sandbox_count: 5
          kind: kind
          console_address: console_address
          route_address: route_address
          oidc: oidc
          created_by: created_by
          supervisor_image: supervisor_image
          service_type: service_type
          tls_mode: tls_mode
          route: route
          database_id: database_id
          name: name
          namespace: namespace
          status: status
        - release_id: release_id
          server_dns_names:
          - server_dns_names
          - server_dns_names
          created_at: 2000-01-23T04:56:07.000+00:00
          credential_driver: credential_driver
          cluster_id: cluster_id
          updated_at: 2000-01-23T04:56:07.000+00:00
          external_reference: external_reference
          id: id
          href: href
          external_dns: external_dns
          phase: phase
          image: image
          active_sandbox_count: 5
          kind: kind
          console_address: console_address
          route_address: route_address
          oidc: oidc
          created_by: created_by
          supervisor_image: supervisor_image
          service_type: service_type
          tls_mode: tls_mode
          route: route
          database_id: database_id
          name: name
          namespace: namespace
          status: status
...
    OpenShellGatewayServiceAccountStatus:
      enum:
      - provisioning
      - ready
      - degraded
      - expired
      - revoking
      - revoked
      - deleting
      - error
      type: string
    OpenShellGatewayServiceAccountCreateRequest:
      additionalProperties: false
      example:
        credential_type: client_secret
        role: openshell-user
        expires_at: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
      properties:
        name:
          maxLength: 128
          minLength: 1
          type: string
        description:
          maxLength: 1024
          nullable: true
          type: string
        credential_type:
          default: client_secret
          enum:
          - client_secret
          type: string
        role:
          $ref: "#/components/schemas/OpenShellGatewayServiceAccountRole"
        expires_at:
          format: date-time
          type: string
      required:
      - name
      type: object
...
    OpenShellGatewayServiceAccountCreateResponse:
      allOf:
      - $ref: "#/components/schemas/OpenShellGatewayServiceAccountListItem"
      - properties:
          credential:
            $ref: "#/components/schemas/OpenShellGatewayServiceAccountCredential"
        required:
        - credential
        type: object
      example:
        role: openshell-user
        subject: subject
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        created_by_user_id: created_by_user_id
        gateway_id: gateway_id
        client_id: client_id
        credential_type: client_secret
        expires_at: 2000-01-23T04:56:07.000+00:00
        revoked_at: 2000-01-23T04:56:07.000+00:00
        updated_at: 2000-01-23T04:56:07.000+00:00
        credential:
          gateway_endpoint: https://openapi-generator.tech
          audience: audience
          grant_type: client_credentials
          client_secret: client_secret
          access_token_lifetime_seconds: 72
          issuer: https://openapi-generator.tech
          client_id: client_id
          gateway_name: gateway_name
          token_endpoint: https://openapi-generator.tech
        name: name
        id: id
        last_error: last_error
        status: provisioning
...
    OpenShellGatewayServiceAccountCapabilities:
      example:
        can_create: true
        allowed_roles:
        - openshell-user
        - openshell-user
        can_manage_all: true
        expiration_policy:
          minimum_seconds: 3600
          default_seconds: 3600
          maximum_seconds: 7262786
      properties:
        can_create:
          type: boolean
        allowed_roles:
          items:
            $ref: "#/components/schemas/OpenShellGatewayServiceAccountRole"
          type: array
        can_manage_all:
          type: boolean
        expiration_policy:
          $ref: "#/components/schemas/OpenShellGatewayServiceAccountExpirationPolicy"
      required:
      - allowed_roles
      - can_create
      - can_manage_all
      - expiration_policy
      type: object
    OpenShellGatewayServiceAccountList:
      example:
        total: 1
        capabilities:
          can_create: true
          allowed_roles:
          - openshell-user
          - openshell-user
          can_manage_all: true
          expiration_policy:
            minimum_seconds: 3600
            default_seconds: 3600
            maximum_seconds: 7262786
        size: 6
        page: 0
        items:
        - role: null
          subject: subject
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          created_by_user_id: created_by_user_id
          gateway_id: gateway_id
          client_id: client_id
          credential_type: client_secret
          expires_at: 2000-01-23T04:56:07.000+00:00
          revoked_at: 2000-01-23T04:56:07.000+00:00
          updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          id: id
          last_error: last_error
          status: provisioning
        - role: null
          subject: subject
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          created_by_user_id: created_by_user_id
          gateway_id: gateway_id
          client_id: client_id
          credential_type: client_secret
          expires_at: 2000-01-23T04:56:07.000+00:00
          revoked_at: 2000-01-23T04:56:07.000+00:00
          updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          id: id
          last_error: last_error
          status: provisioning
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        capabilities:
          $ref: "#/components/schemas/OpenShellGatewayServiceAccountCapabilities"
        items:
          items:
            $ref: "#/components/schemas/OpenShellGatewayServiceAccountListItem"
          type: array
      required:
      - capabilities
      - items
      - page
      - size
      - total
      type: object
...
    RoleList:
      allOf:
      - $ref: "#/components/schemas/List"
      - properties:
          items:
            items:
              $ref: "#/components/schemas/Role"
            type: array
        type: object
      example:
        total: 1
        size: 6
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        page: 0
        id: id
        href: href
        items:
        - updated_at: 2000-01-23T04:56:07.000+00:00
          kind: kind
          permissions: null
          built_in: true
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          description: description
          id: id
          href: href
          display_name: display_name
        - updated_at: 2000-01-23T04:56:07.000+00:00
          kind: kind
          permissions: null
          built_in: true
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          description: description
          id: id
          href: href
          display_name: display_name
...
    GatewayDeletionStatus:
      example:
        deletion_completed_at: 2000-01-23T04:56:07.000+00:00
        deletion_requested_at: 2000-01-23T04:56:07.000+00:00
        external_reference: external_reference
        state: state
        gateway_id: gateway_id
      properties:
        gateway_id:
          type: string
        external_reference:
          type: string
        state:
          description: "active, requested, or completed"
          type: string
        deletion_requested_at:
          format: date-time
          type: string
        deletion_completed_at:
          format: date-time
          type: string
      required:
      - external_reference
      - gateway_id
      - state
      type: object
    GatewayCreateRequest:
      example:
        phase: phase
        release_id: release_id
        image: image
        server_dns_names:
        - server_dns_names
        - server_dns_names
        credential_driver: credential_driver
        oidc: oidc
        supervisor_image: supervisor_image
        cluster_id: cluster_id
        service_type: service_type
        tls_mode: tls_mode
        route: route
        database_id: database_id
        name: name
        external_reference: external_reference
        external_dns: external_dns
        status: status
      properties:
        name:
          type: string
        cluster_id:
          type: string
        release_id:
          type: string
        database_id:
          description: Required placement placeholder; the API server ignores its
            value and assigns the ManagedDatabase
          type: string
        external_reference:
          description: Immutable external resource reference scoped to the authenticated
            caller; repeat creation returns the original gateway
          maxLength: 255
          minLength: 1
          type: string
        external_dns:
          type: string
        tls_mode:
          type: string
        service_type:
          type: string
        status:
          type: string
        phase:
          type: string
        image:
          description: Container image for the gateway deployment
          type: string
        supervisor_image:
          description: Container image for the supervisor sidecar
          type: string
        server_dns_names:
          description: DNS names for TLS certificate SANs
          items:
            type: string
          type: array
        oidc:
          description: JSON-encoded OIDC authentication configuration
          type: string
        route:
          description: JSON-encoded route configuration
          type: string
        credential_driver:
          description: JSON-encoded credential storage driver configuration
          type: string
      required:
      - cluster_id
      - database_id
      - name
      - release_id
      type: object
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http